refactor: Name과 Money의 list 생성 책임을 PlayersFactory로 이동
