+++ src/test/java/domain/gamer/NameTest.java
@@ -4,12 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import util.StringUtil;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class NameTest {
@@ -30,22 +27,9 @@ class NameTest {
 		}).isInstanceOf(IllegalArgumentException.class);
 	}
 
-	@ParameterizedTest
-	@ValueSource(strings = {"allen,kyle,bee", "pobi,jason,cu,woni,brown,jun"})
-	@DisplayName("Name의 list가 제대로 생성되는지 검사")
-	void createTest(String input) {
-		List<Name> names = Name.list(input);
-		List<Name> expectedNames = StringUtil.parseByComma(input)
-				.stream()
-				.map(Name::new)
-				.collect(Collectors.toList());
-
-		assertThat(names).isEqualTo(expectedNames);
-	}
-
 	@Test
 	void namesUnmodifiableTest() {
-		List<Name> names = Name.list("pobi,jason,cu");
+		List<Name> names = PlayersFactory.newNames("pobi,jason,cu");
 
 		assertThatThrownBy(() -> {
 			names.add(new Name("brown"));
